Tabs work!
The set of global variables continues to expand in index.tsx. currentParse.locations (which should not change, unless the user loads another file of course; however, currentParse's blocks may not be up-to-date at any given time with editor content), selectedIndex (a convenience to avoid finding the active tab repeatedly), and listBlocklyWorkplaces (may be renamed; corresponding to the tabs).

Some awful issues with resizing the window... I think blockly workspaces aren't meant to be set hidden quite like this, or something -- need to dig into the documentation more (or it's just undocumented, it seems to be the pattern here).