...Excessively large single update
I'll pretend this commit was so large because of squashing. Anyways, the most recent update is that instructor-defined predicates can be specified (with arg types, although that isn't enforced yet) and appear as blocks on individual toolboxes, following up on adding instructor-defined signatures and getting user-defined variables to work (again, next on the docket is enforcing type checking on those). Together with some simple quantifiers, this is enough for some simple skeleton examples (eg in the Kitteh test file -- I'll upload that once there's more substantive tools).