Type checking
Basic set/variable distinction (which Alloy itself doesn't make, but I think it'd be helpful for student learning). This is accompanied by a renderer change to visually distinguish them. Also, rearranged some code and comments, and added an option to allow or not allow multiple root-level blocks to compile (depending on what the instructor wants).