No instance found popup
Another change to shakudo-integration... maybe they should just be in the same repo, if I'm going to keep making changes to one and wanting them reflected in the other. (Ooh, and maybe build hooks for the java stuff too?) Anyways, there's a popup (of dubious graphical design) for when ALloy runs fine but finds no instances.