selection colors work
this issue has been taunting me for literally years. turns out there's an addon that does bookkeeping for you. frustration/relief. todo: look into if it can be used or modified to handle ordinary row background color in a more sensible way without me needing to be the one bookkeeping.